Moved away from monolithic CMakeLists.txt
- added CMakeLists.txt to each sub-directory level
- moved 3rd party libraries under Extern
- moved 3rd party source to static libraries under Extern
- removed unused SimCoupe.bmp (old SDL application icon)

+++ SDL/OSD.cpp
@@ -60,7 +60,7 @@ std::string OSD::MakeFilePath(PathType type, const std::string& filename)
 
     std::string exe;
     exe.resize(wai_getExecutablePath(nullptr, 0, nullptr));
-    wai_getExecutablePath(exe.data(), exe.length(), nullptr);
+    wai_getExecutablePath(exe.data(), static_cast<int>(exe.length()), nullptr);
     auto exe_path = fs::path(exe).remove_filename();
 
 #if defined(_WINDOWS)
